special_requirements.txt
# Special requirements.txt for Python 3.13
# Updated package versions for compatibility with Python 3.13

# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
aiofiles==24.1.0

# Adafruit libraries
Adafruit-Blinka==8.47.0
adafruit-circuitpython-ads1x15==2.2.26
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
Adafruit-GPIO==1.0.3

# OLED display
luma.core==2.4.2
luma.oled==3.13.0

# GPIO (Note: RPi.GPIO may have limited Python 3.13 support, consider lgpio as alternative)
# RPi.GPIO==0.7.1  # May not be compatible with Python 3.13
lgpio==0.2.2.0  # Alternative GPIO library with Python 3.13 support
gpiozero==2.0.1  # High-level GPIO library, Python 3.13 compatible

# Serial communication
pyserial==3.5

# SPI/I2C
smbus2==0.4.3
spidev==3.6

# Scientific computing (Python 3.13 compatible versions)
numpy==2.1.0
scipy==1.14.1
matplotlib==3.9.2

# USB and communication
pyusb==1.2.1
pyftdi==0.55.4

# WS281x LED strip (may need compilation for Python 3.13)
rpi_ws281x==5.0.0

# Security
cryptography==43.0.1
bcrypt==4.2.0
paramiko==3.5.0
PyNaCl==1.5.0

# Utilities
click==8.1.7
colour==0.1.5
pillow==10.4.0
python-dateutil==2.9.0
packaging==24.1
six==1.16.0

# Additional dependencies
cffi==1.17.1
pycparser==2.22
pyparsing==3.1.4
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
contourpy==1.3.0

# Note: If RPi.GPIO doesn't work with Python 3.13, you'll need to update your code to use lgpio or gpiozero instead
# Migration guide:
# - lgpio is a direct replacement for RPi.GPIO with similar API
# - gpiozero provides a higher-level, more Pythonic interface
